+++ packages/config/src/api/ConfigSecure.ts
@@ -9,6 +9,7 @@
 *
 */
 
+import * as fs from "fs";
 import * as JSONC from "comment-json";
 import * as lodash from "lodash";
 import { ConfigApi } from "./ConfigApi";
@@ -261,6 +262,21 @@ export class ConfigSecure extends ConfigApi {
         }
 
         return { path: securePath, prop: secureProp };
+    }
+
+    /**
+     * Delete secure properties stored for team config files that do not exist.
+     * @returns Array of file paths for which properties were deleted
+     */
+    public rmUnusedProps(): string[] {
+        const prunedFiles: string[] = [];
+        for (const filename of Object.keys(this.mConfig.mSecure)) {
+            if (!fs.existsSync(filename)) {
+                delete this.mConfig.mSecure[filename];
+                prunedFiles.push(filename);
+            }
+        }
+        return prunedFiles;
     }
 
     /**